Moving tests directory and updating imports.
