+++ src/azure_pdf_parser/__init__.py
@@ -1,2 +1,4 @@
 from .azure_wrapper import AzureApiWrapper
 from .convert import azure_api_response_to_parser_output
+from .experimental_base import ExperimentalParserOutput
+from .base import PDFPage
