+++ src/azure_pdf_parser/convert.py
@@ -14,7 +14,7 @@ from cpr_data_access.parser_models import (
     ParserInput,
     CONTENT_TYPE_PDF,
 )
-from azure_pdf_parser.experimental_base import (
+from .experimental_base import (
     ExperimentalTableCell,
     ExperimentalBoundingRegion,
     ExperimentalPDFTableBlock,
